Move ElapseTimer out of SmingCore and into Services/Profiling
